backend/app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-cloud-bigquery==3.25.0
google-auth==2.33.0
pydantic==2.9.2
email-validator==2.2.0
python-dotenv==1.0.1